refactor(bspline): move _find_knot_index_and_multiplicity to _bspline_knots
This function is a general knot-vector query utility, so it belongs
in _bspline_knots.py alongside the other knot helpers rather than
in the knot-removal module.
